refactor: TestDataBuilder 패키지 변경 및 static import, 인증 통합 테스트 리팩토링
