+++ src/test/java/com/jnulocker/events/utils/EventTestUtil.java
@@ -1,5 +1,9 @@
 package com.jnulocker.events.utils;
 
+import static events.domain.EventTestDataBuilder.*;
+import static organization.domain.DepartmentTestDataBuilder.*;
+import static organization.domain.OrganizationTestDataBuilder.*;
+
 import com.jnulocker.events.adapter.out.EventRepository;
 import com.jnulocker.events.adapter.out.FloorRepository;
 import com.jnulocker.events.adapter.out.LockerRepository;
@@ -10,13 +14,10 @@ import com.jnulocker.organization.adapter.out.DepartmentRepository;
 import com.jnulocker.organization.adapter.out.OrganizationRepository;
 import com.jnulocker.organization.domain.Department;
 import com.jnulocker.organization.domain.Organization;
-import events.builder.EventTestDataBuilder;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import organization.builder.DepartmentTestDataBuilder;
-import organization.builder.OrganizationTestDataBuilder;
 
 @Component
 public class EventTestUtil {
@@ -33,16 +34,15 @@ public class EventTestUtil {
 
     public Event createEventWithFloorAndLockers(List<Integer> lockersPerFloor) {
         // 조직 생성 및 저장
-        Organization organization = OrganizationTestDataBuilder.builder().build();
+        Organization organization = organizationBuilder().build();
         Organization savedOrganization = organizationRepository.save(organization);
 
         // 학과 생성 및 저장
-        Department department =
-                DepartmentTestDataBuilder.builder().withOrganization(savedOrganization).build();
+        Department department = departmentBuilder().withOrganization(savedOrganization).build();
         Department savedDepartment = departmentRepository.save(department);
 
         // 이벤트 생성 및 저장
-        Event event = EventTestDataBuilder.builder().withDepartment(savedDepartment).build();
+        Event event = eventBuilder().withDepartment(savedDepartment).build();
         Event savedEvent = eventRepository.save(event);
 
         // 층 생성 및 저장
